Move missing files to observer dir
