+++ setup.cfg
@@ -41,6 +41,6 @@ node_observer =
   tail-text = tei_transform.observer.tail_text_observer:TailTextObserver
   p-div-sibling = tei_transform.observer.p_as_div_sibling_observer:PAsDivSiblingObserver
   div-text = tei_transform.observer.div_text_observer:DivTextObserver
-  double-item = tei_transform.double_item_observer:DoubleItemObserver
-  list-div-sibling = tei_transform.list_as_div_sibling_observer:ListAsDivSiblingObserver
-  double-cell = tei_transform.double_cell_observer:DoubleCellObserver
+  double-item = tei_transform.observer.double_item_observer:DoubleItemObserver
+  list-div-sibling = tei_transform.observer.list_as_div_sibling_observer:ListAsDivSiblingObserver
+  double-cell = tei_transform.observer.double_cell_observer:DoubleCellObserver
