+++ tei_transform/observer/__init__.py
@@ -1,5 +1,6 @@
 from tei_transform.observer.classcode_observer import ClasscodeObserver
 from tei_transform.observer.div_text_observer import DivTextObserver
+from tei_transform.observer.double_cell_observer import DoubleCellObserver
 from tei_transform.observer.double_item_observer import DoubleItemObserver
 from tei_transform.observer.filename_element_observer import FilenameElementObserver
 from tei_transform.observer.head_after_p_element_observer import (
@@ -7,6 +8,7 @@ from tei_transform.observer.head_after_p_element_observer import (
 )
 from tei_transform.observer.head_with_type_attr_observer import HeadWithTypeAttrObserver
 from tei_transform.observer.id_attribute_observer import IdAttributeObserver
+from tei_transform.observer.list_as_div_sibling_observer import ListAsDivSiblingObserver
 from tei_transform.observer.notesstmt_observer import NotesStmtObserver
 from tei_transform.observer.p_as_div_sibling_observer import PAsDivSiblingObserver
 from tei_transform.observer.schemalocation_observer import SchemaLocationObserver
